requirements.txt
appdirs==1.4.3
asn1crypto==1.4.0
azure-common==1.1.25
azure-core==1.8.2
azure-storage-blob==12.5.0
boto3==1.15.18
botocore==1.18.18
CacheControl==0.12.6
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
cryptography==2.9.2
distlib==0.3.0
distro==1.4.0
html5lib==1.0.1
idna==2.10
ipaddr==2.2.0
isodate==0.6.0
jmespath==0.10.0
lockfile==0.12.2
msgpack==0.6.2
msrest==0.6.19
numpy==1.20.2
oauthlib==3.1.0
oscrypto==1.2.1
packaging==20.3
pandas==1.2.3
patsy==0.5.1
pep517==0.8.2
progress==1.5
pyarrow==3.0.0
pycparser==2.20
pycryptodomex==3.9.8
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.6
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
s3transfer==0.3.3
scipy==1.6.2
six==1.15.0
slack-sdk==3.4.2
snowflake-connector-python==2.4.1
statsmodels==0.12.2
urllib3==1.25.11
webencodings==0.5.1
